Changed sample period to seconds (instead of PMU cycles)

+++ analysis_utility/analysis_tool/src/config/config.cpp
@@ -44,6 +44,6 @@ unsigned Config::window;
 unsigned Config::sdsocVersion;
 QString Config::extraCompileOptions;
 QString Config::projectDir;
-int64_t Config::overrideSamplePeriod;
+double Config::overrideSamplePeriod;
 bool Config::overrideSamplePc;
 bool Config::overrideNoSamplePc;
